requirements.txt
# Deep Learning Core
torch>=2.0.0
torchvision>=0.15.0

# Data Computation and Arrays
numpy>=1.24.0

# Image Processing and Remote Sensing
Pillow>=9.5.0
tifffile>=2023.4.12
opencv-python-headless>=4.8.0

# Visualization
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0